requirements.txt
Django~=4.1.2
djangorestframework
django-reversion
django-filter
celery